requirements.txt
# LangChain Core
langchain
langchain-core

# OpenAI Integration
langchain-openai
openai

# Anthropic Integration
langchain-anthropic

# Google Gemini (PaLM) Integration
langchain-google-genai
google-generativeai

# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub

# Environment Variable Management
python-dotenv

# Machine Learning Utilities
numpy
scikit-learn

#vectorstore
chromadb==0.4.22
#pdf
pypdf==4.0.1

#ui
streamlit==1.32.0

#community
langchain-community==0.0.24

# Embeddings and Sentence Transformers -hf
sentence-transformers